Minor logic optimisations
- Replaced local variable lists with local variable literal tuples.
- Pre-calculated the card keys tuples to avoid string formatting in
logic rules.

In terms of performance, from most performant to least performant is
generally as follows:
1) Inlined literal tuple
1.1) Global pre-created tuple (literal or non-literal)
1.1) Global pre-created list
1.2) Literal tuple assigned to a local variable

2) A new non-literal tuple
2.1) A new list

The difference between the best and worst of these is rather small, so
it is unlikely to see any noticeable difference from replacing newly
created lists with global lists or tuple literals, unless the number of
new lists being created was huge.

---

In general, tuples are a tiny bit more memory efficient because Python
reserves some memory for tuples and because a tuple is a fixed size, the
memory that it needs to be assigned is also fixed, whereas lists assign
a little extra memory to make .append() faster.

A literal tuple, a tuple containing only literals (can include other
literal tuples) is written directly into the Python bytecode so is
usually the fastest.

Assigning a literal tuple to a local variable requires a write into the
locals(), and then using that local variable again later requires
getting that variable from the locals(). Python, unlike most compiled
programming languages is not able to automatically inline simple uses,
so a manually inlined literal tuple is slightly faster because it skips
the 'set' into and 'get' from the locals().

Getting a tuple (or list) from a global variable requires a 'get' like
getting a local variable, but skips the 'set' because the global
variable has been defined in advance.

Creating a new non-literal tuple or a new list has to create that new
tuple/list object, set the items into the tuple/list and only then can
the tuple/list be used.