Fix trivial bug in deleting unreachable evolution/dexsanity locations
The second argument to state.can_reach() is not the player, it is a hint
towards what kind of spot should have its reachability checked, used
when the first argument is a string.

Because the first argument was a Location, the second, incorrect,
argument was never being used, so this bug had no effect on the world.

Generally, there is not much point to using state.can_reach() when the
location/entrance/region has already been acquired, so these uses of
state.can_reach() have been replaced with location.can_reach().